Refactor mask
