+++ src/mask/bitmask.c
