+++ src/mask/mask.pyi
