+++ src/mask/mask_doc.h
